"Updated schema definitions for various tables, including AuthorCommision, Book, BookStore, Buyer, Order, OrderItem, and Seller, and created new tables for many-to-many relationships."

+++ lib/phx_book_store/orders/order.ex
@@ -5,7 +5,7 @@ defmodule PhxBookStore.Orders.Order do
 
   schema "orders" do
     field :status, :string
-    field :total_amount, :float
+    field :total_amount, :integer
     field :currency, :string
     field :stripe_charge_id, :string
     field :stripe_payment_intent_id, :string
